10/7

+++ 03_Django/04_django_crue_review/articles/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from django.core.exceptions import ValidationError
 from .models import Article
+from IPython import embed
+
 def index(request):
     articles = Article.objects.order_by('-pk')
     # print(articles)
@@ -9,24 +11,33 @@ def index(request):
     return render(request, 'articles/index.html', context)
 # Create your views here.
 
-def new(request):
-    return render(request, 'articles/new.html')
+# def new(request):
+#     embed()
+#     return render(request, 'articles/new.html')
 
 def create(request):
-    try:
+    if request.method == 'POST':
         title = request.POST.get('title')
         content = request.POST.get('content')
         article = Article(title=title, content=content)
         article.full_clean()
-    except ValidationError:
-        raise ValidationError('Your Error Message')
-    else:
         article.save()
         return redirect('articles:detail', article.pk)
+    else:
+        return render(request, 'articles/create.html')
+
+    # try:
+    #     title = request.POST.get('title')
+    #     content = request.POST.get('content')
+    #     article = Article(title=title, content=content)
+    #     article.full_clean()
+    # except ValidationError:
+    #     raise ValidationError('Your Error Message')
+    # else:
+    #     article.save()
+    #     return redirect('articles:detail', article.pk)
 
 
-    # title = request.POST.get('title')
-    # content = request.POST.get('content')
 
     # 1번째 방법
     # article = Article()
@@ -52,17 +63,26 @@ def detail(request, pk):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
+    if request.method == "POST":
+        article.delete()
+        return redirect('articles:index')
+    else:
+        context = {'article': article}
+        return render(request, 'articles/detail.html', context)
 
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {'article': article}
-    return render(request, 'articles/edit.html', context)
+# def edit(request, pk):
+#     article = Article.objects.get(pk=pk)
+#     context = {'article': article}
+#     return render(request, 'articles/edit.html', context)
 
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-    return redirect('articles:detail', article.pk)
+    if request.method == 'GET':
+        context = {'article': article}
+        return render(request, 'articles/edit.html', context)
+    else:
+        # embed()
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect('articles:detail', article.pk)
